requirements.txt
annotated-types==0.7.0
anyio==4.11.0
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.3.0
decorator==5.2.1
distro==1.9.0
fastapi==0.117.1
future==1.0.0
geocoder==1.38.1
geographiclib==2.1
geopy==2.4.1
greenlet==3.2.4
gunicorn==23.0.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
jiter==0.11.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-core==0.3.76
langchain-openai==0.3.33
langchain-text-splitters==0.3.11
langgraph==0.6.7
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.4
langgraph-sdk==0.2.9
langsmith==0.4.31
openai==1.109.1
orjson==3.11.3
ormsgpack==1.10.0
packaging==25.0
psycopg2-binary==2.9.10
pydantic==2.11.9
pydantic_core==2.33.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.3
ratelim==0.1.6
regex==2025.9.18
requests==2.32.5
requests-toolbelt==1.0.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.43
starlette==0.48.0
tenacity==9.1.2
tiktoken==0.11.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.15.0
urllib3==2.5.0
uvicorn==0.37.0
xxhash==3.5.0
zstandard==0.25.0
alembic==<latest>
SQLAlchemy==2.0.43

